// src/types/enums/taskType.ts
export enum TaskType {
  SubmitScan = 0,
  PrimeShields = 1,
  FuelEngines = 2,
  ChartCourse = 3,
  StartReactor = 4,
  SwipeCard = 5,
  ClearAsteroids = 6,
  UploadData = 7,
  InspectSample = 8,
  EmptyChute = 9,
  EmptyGarbage = 10,
  AlignEngineOutput = 11,
  FixWiring = 12,
  CalibrateDistributor = 13,
  DivertPower = 14,
  UnlockManifolds = 15,
  ResetReactor = 16,
  FixLights = 17,
  CleanO2Filter = 18,
  FixComms = 19,
  RestoreO2 = 20,
  StabilizeSteering = 21,
  AssembleArtifact = 22,
  SortSamples = 23,
  MeasureWeather = 24,
  EnterIdCode = 25,
  BuyBeverage = 26,
  ProcessData = 27,
  RunDiagnostics = 28,
  WaterPlants = 29,
  MonitorO2 = 30,
  StoreArtifacts = 31,
  FillCanisters = 32,
  ActivateWeatherNodes = 33,
  InsertKeys = 34,
  ResetSeismic = 35,
  ScanBoardingPass = 36,
  OpenWaterways = 37,
  ReplaceWaterJug = 38,
  RepairDrill = 39,
  AlignTelescope = 40,
  RecordTemperature = 41,
  RebootWifi = 42,
  PolishRuby = 43,
  ResetBreakers = 44,
  Decontaminate = 45,
  MakeBurger = 46,
  UnlockSafe = 47,
  SortRecords = 48,
  PutAwayPistols = 49,
  FixShower = 50,
  CleanToilet = 51,
  DressMannequin = 52,
  PickUpTowels = 53,
  RewindTapes = 54,
  StartFans = 55,
  DevelopPhotos = 56,
  GetBiggolSword = 57,
  PutAwayRifles = 58,
  StopCharles = 59,
}